Move sessions to context
